+++ apps/router-tester/app/split-view/[type]/index.tsx
@@ -2,7 +2,7 @@ import { PlatformColor, Text, View } from 'react-native';
 
 export default function Index() {
   return (
-    <View style={{ flex: 1, justifyContent: 'center', alignItems: 'center' }}>
+    <View style={{ padding: 24 }}>
       <Text style={{ color: PlatformColor('label'), fontSize: 24, fontWeight: 'bold' }}>
         Nothing is selected
       </Text>